% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{create_professional_map}
\alias{create_professional_map}
\title{Crée une carte ggplot stylisée pour une variable régionale}
\usage{
create_professional_map(
  spatial_data,
  var,
  title = "",
  palette = "viridis",
  legend_title = ""
)
}
\arguments{
\item{spatial_data}{Données spatiales (avec géométrie + variable à représenter)}

\item{var}{Nom de la variable à représenter (ex: "QuantiteConsommeeKG")}

\item{title}{Titre de la carte}

\item{palette}{Palette de couleurs viridis (ex: "viridis", "plasma", "magma", "cividis")}

\item{legend_title}{Titre de la légende}
}
\value{
Un objet ggplot2
}
\description{
Crée une carte ggplot stylisée pour une variable régionale
}
