% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{generate_interactive_map}
\alias{generate_interactive_map}
\title{Génère une carte interactive à partir de données de consommation}
\usage{
generate_interactive_map(
  base_data,
  shapefile_path = NULL,
  variable = "QuantiteConsommeeKG",
  popup_vars = c("region", "ValeurConsommee", "CaloriesParTete"),
  popup_labels = c("Région", "Valeur (F CFA)", "Calories Moyennes"),
  palette = "YlOrRd",
  legend_title = "Valeur"
)
}
\arguments{
\item{base_data}{Données brutes avec colonnes `region`}

\item{shapefile_path}{Chemin vers un shapefile personnalisé}

\item{variable}{Variable à représenter}

\item{popup_vars}{Colonnes pour popup}

\item{popup_labels}{Libellés pour popup}

\item{palette}{Palette de couleurs}

\item{legend_title}{Titre de la légende}
}
\value{
Un objet leaflet
}
\description{
Génère une carte interactive à partir de données de consommation
}
