% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{generate_spatial_map}
\alias{generate_spatial_map}
\title{Génère directement une carte spatiale stylisée}
\usage{
generate_spatial_map(
  base_data,
  shapefile_path = NULL,
  variable = "QuantiteConsommeeKG",
  palette = "viridis",
  title = "",
  legend_title = ""
)
}
\arguments{
\item{base_data}{Base de consommation brute}

\item{shapefile_path}{Chemin vers le shapefile}

\item{variable}{Variable à cartographier}

\item{palette}{Palette de couleurs viridis}

\item{title}{Titre de la carte}

\item{legend_title}{Titre de la légende}
}
\value{
Un ggplot2
}
\description{
Génère directement une carte spatiale stylisée
}
