% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{integrate_data_with_shapes}
\alias{integrate_data_with_shapes}
\title{Intègre les données de consommation avec les shapes régionaux}
\usage{
integrate_data_with_shapes(consumption_data, region_sf)
}
\arguments{
\item{consumption_data}{Données de consommation (data.frame avec une colonne "region")}

\item{region_sf}{Objet sf des régions (de load_regional_shapes)}
}
\value{
Un sf dataframe enrichi
}
\description{
Intègre les données de consommation avec les shapes régionaux
}
