% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_regional_shapes}
\alias{load_regional_shapes}
\title{Charger les limites régionales du Sénégal}
\usage{
load_regional_shapes(path = NULL)
}
\arguments{
\item{path}{Chemin vers le shapefile. Si NULL, utilise celui inclus dans le package.}
}
\value{
Un objet sf avec les régions
}
\description{
Charger les limites régionales du Sénégal
}
