% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_packages}
\alias{require_packages}
\title{Charge une liste de packages requis}
\usage{
require_packages(packages)
}
\arguments{
\item{packages}{Vecteur de noms de packages à charger}
}
\value{
Charge les packages s’ils sont installés, ou affiche un message d’erreur clair
}
\description{
Charge une liste de packages requis
}
