% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{create_interactive_map}
\alias{create_interactive_map}
\title{Crée une carte interactive Leaflet pour visualiser une variable régionale}
\usage{
create_interactive_map(
  spatial_data,
  var,
  popup_vars = NULL,
  palette = "YlOrRd"
)
}
\arguments{
\item{spatial_data}{Données spatialisées (objet `sf` avec géométrie)}

\item{var}{Nom de la variable numérique à cartographier}

\item{popup_vars}{Variables à afficher dans le popup (par défaut : nom de la région + var)}

\item{palette}{Palette de couleurs (ex: "YlOrRd", "Viridis", "Blues", etc.)}
}
\value{
Un objet leaflet
}
\description{
Crée une carte interactive Leaflet pour visualiser une variable régionale
}
